查看首页、查看活动列表页、翻页均不需要登录。登录窗口增加默认账户提示

+++ src/lib/features/activity/api/activity-api.ts
@@ -2,6 +2,8 @@ import { toActivityVO } from "@/lib/features/activity/shared/activity.ts";
 import { Hono } from "hono";
 import { deleteById, findByCondition, getObjById, updateObj } from "@/lib/features/activity/service/activity-service.ts";
 import { activityConditionCheck, editActivityCheck, activityById, activityPageResponseSchema } from "../shared/activity-check";
+import { ApplicationException } from "@/lib/types/application-exception.ts";
+import { COMMON_RESPONSE } from "@/lib/types/error-type.ts";
 
 const app = new Hono();
 export const activityApi = app
@@ -24,12 +26,18 @@ export const activityApi = app
     })
     .put('/updateObj', async (context) => {
         const currentUser = context.get('authz');
+        if (!currentUser) {
+            throw new ApplicationException(COMMON_RESPONSE.UNAUTHORIZED);
+        }
         const activity = editActivityCheck.parse(await context.req.json());
         const success = await updateObj(activity, currentUser);
         return context.json(success);
     })
     .delete('/deleteById', async (context) => {
         const currentUser = context.get('authz');
+        if (!currentUser) {
+            throw new ApplicationException(COMMON_RESPONSE.UNAUTHORIZED);
+        }
         const { id } = activityById.parse(context.req.query());
         const success = await deleteById(id, currentUser);
         return context.json(success);
